server/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
aiodocker==0.25.0
websockets==16.0
httpx==0.28.1
aiofiles==25.1.0
jinja2==3.1.6
py7zr==1.1.2
bcrypt==5.0.0
pyyaml==6.0.2
